completed exercise 3
It was interesting to learn that you can work even with private function, I'm starting to understand it better but still want to read up on the terminology, I'm convinced that the more I will work with it the better it will go. It's starting to click but still a long way to go. At first I was like buh PHP, but with these exercises I'm starting to like it more and more