writing my README
I want to write down everything down I did, the information I looked up in a proper way. I don't think I will be able to finish it today because it was not so smart of me to not add anything to my readme these past 2 days. But my girlfriend has to work during the weekend so I will do my best to finish it by then.